+++ lib/jsonable.dart
@@ -1,10 +1,11 @@
 library jsonable;
 
-import 'package:jsonable/src/jsonable.dart';
 import 'package:jsonable/src/validator/exceptions.dart';
 import 'package:jsonable/src/validator/rules.dart';
-export "./src/jsonable.dart" show Jsonable;
-export "./exceptions.dart";
+import "./src/mixin_jsonable.dart" show Jsonable;
+
+export "./src/mixin_jsonable.dart" show Jsonable;
+export "./src/validator/exceptions.dart";
 export "./src/validator/rules.dart" show Rules;
 
 typedef JsonableBuilder<E extends Jsonable> = E Function();
@@ -25,8 +26,11 @@ abstract class JType<E> {
   /// this is Keyname in jsonable object
   String keyname;
 
-  E get get => this._value;
-  void set set(E value) => this._value = value;
+  E get value => this._value;
+  void set value(E value) => this._value = value;
+
+  E getValue() => this._value;
+  void setValue(E value) => this._value = value;
 
   List<RuleException> validate() {
     return _rules
@@ -40,34 +44,53 @@ abstract class JString extends JType<String>
     implements Comparable<String>, Pattern {
   JString({String initialValue, List<Rule> rules: const []})
       : super(value: initialValue, rules: rules);
+
+  String getString() => this._value;
+  void setString(String value) => this._value = value;
 }
 
 abstract class JNum extends JType<num> implements Comparable<num> {
   JNum({num initialValue, List<Rule> rules})
       : super(value: initialValue, rules: rules);
+
+  num getNum() => this._value;
+  void setNum(num value) => this._value = value;
 }
 
 abstract class JBool extends JType<bool> {
   JBool({bool initialValue, List<Rule> rules})
       : super(value: initialValue, rules: rules);
+  bool getBool() => this._value;
+  void setBool(bool value) => this._value = value;
 }
 
 abstract class JList<E> extends JType<List<E>> implements List<E> {
   JList({List<E> initialValue, List<Rule> rules})
       : super(value: initialValue, rules: rules);
+
+  List<E> getList() => this._value;
+  void setList(List<E> value) => this._value = value;
 }
 
 abstract class JClass<E extends Jsonable> extends JType<E> {
   JClass({Jsonable initialValue, List<Rule> rules})
       : super(value: initialValue, rules: rules);
+  E getClass() => this._value;
+  void setClass(E value) => this._value = value;
 }
 
 abstract class JMap<E, R> extends JType<Map> implements Map<E, R> {
   JMap({Map initialValue, List<Rule> rules})
       : super(value: initialValue, rules: rules);
+
+  Map<E, R> getMap() => this._value;
+  void setMap(Map<E, R> value) => this._value = value;
 }
 
 abstract class JDynamic extends JType<dynamic> {
   JDynamic({dynamic initialValue, List<Rule> rules})
       : super(value: initialValue, rules: rules);
+
+  dynamic getDynamic() => this._value;
+  void setDynamic(dynamic value) => this._value = value;
 }