+++ test/jamp_test.dart
@@ -17,12 +17,12 @@ main() {
     var j = Person();
     j.fromMap(m);
     test("tags test", () {
-      expect(j.tags.get["1"]["map"], "map");
-      expect(j.tags.get["2"]["test"], "test");
+      expect(j.tags.value["1"]["map"], "map");
+      expect(j.tags.value["2"]["test"], "test");
     });
     test("contacts", () {
-      expect(j.contacts["mario"].name.get, "Mario");
-      expect(j.contacts["super"].name.get, "super");
+      expect(j.contacts["mario"].name.value, "Mario");
+      expect(j.contacts["super"].name.value, "super");
     });
   });
 }