+++ test/simple_class_test.dart
@@ -15,18 +15,18 @@ main() {
     group("constructor value", () {
       var p = Person(name: "nico", surname: "spina");
       test("", () {
-        expect(p.name.get, "nico");
+        expect(p.name.value, "nico");
       });
       test("", () {
-        expect(p.surname.get, "spina");
+        expect(p.surname.value, "spina");
       });
     });
     group("toMap()", () {
       var p = Person(name: "Nico", surname: "Spina");
       var mapped = p.toMap();
       test("value", () {
-        expect(p.name.get, mapped["name"]);
-        expect(p.surname.get, mapped["surname"]);
+        expect(p.name.value, mapped["name"]);
+        expect(p.surname.value, mapped["surname"]);
       });
     });
     group("fromMap", () {
@@ -34,8 +34,8 @@ main() {
       var p = Person();
       p.fromMap(mapped);
       test("value", () {
-        expect(p.name.get, mapped["name"]);
-        expect(p.surname.get, mapped["surname"]);
+        expect(p.name.value, mapped["name"]);
+        expect(p.surname.value, mapped["surname"]);
       });
     });
   });