+++ CHANGELOG.md
@@ -5,8 +5,12 @@
 * **New** add support for validation field of Jsonable object (see readme)
 * **Breaking change** delete support Jsonable with reflection use [dson](https://pub.dev/packages/dson)
 * **Breaking Change** setter and getter value have been removed and replaced with
-    * void set set(E value) - set value within JType
-    * E get get; -  get value in JType
+    * void setValue(E value) - set value within JType
+    * E getValue; -  get value in JType
+    **every JType**
+    * setString(String value)
+    * setNum(num value)
+    * setBool(bool value)
 * **Breaking Change** removed (experimental) operator "<<" and operator ">>"